rewrite resovling schema diffs, move functions out of page comp
